requirements.txt
asgiref==3.7.2
greenlet==3.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
phonenumbers==8.13.26
phonenumberslite==8.13.26
django_phonenumber_field==7.2.0
faker==20.1.0
